stock-analyst-api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
yfinance>=0.2.40
feedparser>=6.0.11
langgraph>=0.2.20
langchain-openai>=0.2.0
langchain-core>=0.3.0
